Moved ServerInertial device from include folder to libYARP_dev/src/modules like the others and split it into .h and .cpp
